+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/MoshiChatApi.kt
@@ -117,6 +117,7 @@ import io.getstream.chat.android.models.MemberData
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.models.Mute
 import io.getstream.chat.android.models.Option
+import io.getstream.chat.android.models.PendingMessage
 import io.getstream.chat.android.models.Poll
 import io.getstream.chat.android.models.PollConfig
 import io.getstream.chat.android.models.QueryDraftsResult
@@ -340,6 +341,10 @@ constructor(
         ).mapDomain { response ->
             response.message.toDomain()
         }
+    }
+
+    override fun getPendingMessage(messageId: String): Call<PendingMessage> {
+        return messageApi.getMessage(messageId).mapDomain { it.toDomain() }
     }
 
     override fun deleteMessage(messageId: String, hard: Boolean): Call<Message> {
@@ -947,6 +952,7 @@ constructor(
                 messages = response.messages.map {
                     it.toDomain(channelInfo).enrichWithCid(channel.cid)
                 },
+                pendingMessages = response.pending_messages.map { it.toDomain() },
                 pinnedMessages = response.pinned_messages.map {
                     it.toDomain(channelInfo).enrichWithCid(channel.cid)
                 },