+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/DomainMapping.kt
@@ -35,6 +35,7 @@ import io.getstream.chat.android.client.api2.model.dto.DownstreamModerationDetai
 import io.getstream.chat.android.client.api2.model.dto.DownstreamModerationDto
 import io.getstream.chat.android.client.api2.model.dto.DownstreamMuteDto
 import io.getstream.chat.android.client.api2.model.dto.DownstreamOptionDto
+import io.getstream.chat.android.client.api2.model.dto.DownstreamPendingMessageDto
 import io.getstream.chat.android.client.api2.model.dto.DownstreamPollDto
 import io.getstream.chat.android.client.api2.model.dto.DownstreamReactionDto
 import io.getstream.chat.android.client.api2.model.dto.DownstreamReactionGroupDto
@@ -53,6 +54,7 @@ import io.getstream.chat.android.client.api2.model.response.AppSettingsResponse
 import io.getstream.chat.android.client.api2.model.response.BannedUserResponse
 import io.getstream.chat.android.client.api2.model.response.BlockUserResponse
 import io.getstream.chat.android.client.api2.model.response.FileUploadConfigDto
+import io.getstream.chat.android.client.api2.model.response.MessageResponse
 import io.getstream.chat.android.client.extensions.syncUnreadCountWithReads
 import io.getstream.chat.android.core.internal.StreamHandsOff
 import io.getstream.chat.android.models.Answer
@@ -80,6 +82,7 @@ import io.getstream.chat.android.models.Moderation
 import io.getstream.chat.android.models.ModerationAction
 import io.getstream.chat.android.models.Mute
 import io.getstream.chat.android.models.Option
+import io.getstream.chat.android.models.PendingMessage
 import io.getstream.chat.android.models.Poll
 import io.getstream.chat.android.models.PushProvider
 import io.getstream.chat.android.models.Reaction
@@ -231,6 +234,24 @@ internal class DomainMapping(
             silent = message.silent,
             text = message.text,
             extraData = message.extraData ?: emptyMap(),
+        )
+
+    /**
+     * Transforms [DownstreamPendingMessageDto] to [PendingMessage].
+     */
+    internal fun DownstreamPendingMessageDto.toDomain(): PendingMessage =
+        PendingMessage(
+            message = message.toDomain(),
+            metadata = metadata.orEmpty(),
+        )
+
+    /**
+     * Transforms [MessageResponse] to [PendingMessage].
+     */
+    internal fun MessageResponse.toDomain(): PendingMessage =
+        PendingMessage(
+            message = message.toDomain(),
+            metadata = pending_message_metadata.orEmpty(),
         )
 
     /**
@@ -555,6 +576,7 @@ internal class DomainMapping(
         automodBehavior = automod_behavior,
         blocklistBehavior = blocklist_behavior ?: "",
         commands = commands.map { it.toDomain() },
+        markMessagesPending = mark_messages_pending,
     )
 
     /**